ChatBot Using Existing LLM/requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
faiss-cpu
langchain_google_genai
beautifulsoup4
requests
